Better code separation and PGN, MoveForm added
Changes:

- Add a Fen class responsible for parsing Forsyth-Edwards Notation and
returning an array of spaces that each represent one square on the chess
board
- Add a Move class that has a bunch of static methods that figure out
how pieces are allowed to move. It's primary API is the possibleMoves
method which takes an array of spaces and returns an array of the same
size, where each member is the number of possible valid moves ending on
the corresponding space (this is basically the heat map portion).
- Add a Pgn class for producing a new board state from a given board
state and a PGN move (something like ka6)
- Add a MoveForm component for testing out the PGN logic.

Next steps:

- Add in Redux and restructure the handleMove code to use redux
- The global state should include more than just the FEN and heatmap, it
should also include things like who is the next to move, castling
options, etc. (basically everything a FEN includes).